+++ apps/workflows-dashboard/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @ballerine/workflows-dashboard
 
+## 0.2.17
+
+### Patch Changes
+
+- Bump
+- Updated dependencies
+  - @ballerine/common@0.9.33
+  - @ballerine/ui@0.5.33
+
+## 0.2.16
+
+### Patch Changes
+
+- d
+- Updated dependencies
+  - @ballerine/common@0.9.32
+  - @ballerine/ui@0.5.31
+
+## 0.2.15
+
+### Patch Changes
+
+- version bump
+- Updated dependencies
+  - @ballerine/common@0.9.31
+  - @ballerine/ui@0.5.30
+
+## 0.2.14
+
+### Patch Changes
+
+- Bump
+- Updated dependencies
+  - @ballerine/common@0.9.30
+  - @ballerine/ui@0.5.29
+
 ## 0.2.13
 
 ### Patch Changes