+++ examples/headless-example/CHANGELOG.md
@@ -1,9 +1,37 @@
 # @ballerine/headless-example
 
+## 0.3.43
+
+### Patch Changes
+
+- Bump
+- Updated dependencies
+  - @ballerine/common@0.9.33
+  - @ballerine/workflow-browser-sdk@0.6.44
+
+## 0.3.42
+
+### Patch Changes
+
+- d
+- Updated dependencies
+  - @ballerine/common@0.9.32
+  - @ballerine/workflow-browser-sdk@0.6.43
+
+## 0.3.41
+
+### Patch Changes
+
+- version bump
+- Updated dependencies
+  - @ballerine/common@0.9.31
+  - @ballerine/workflow-browser-sdk@0.6.42
+
 ## 0.3.40
 
 ### Patch Changes
 
+- Bump
 - Updated dependencies
   - @ballerine/common@0.9.30
   - @ballerine/workflow-browser-sdk@0.6.41