+++ packages/common/CHANGELOG.md
@@ -1,16 +1,34 @@
 # @ballerine/common
 
+## 0.9.33
+
+### Patch Changes
+
+- Bump
+
+## 0.9.32
+
+### Patch Changes
+
+- d
+
+## 0.9.31
+
+### Patch Changes
+
+- version bump
+
 ## 0.9.30
 
 ### Patch Changes
 
-- package sync
+- Bump
 
 ## 0.9.29
 
 ### Patch Changes
 
-- Added signing logic with crypt
+- version bump fix
 
 ## 0.9.28