+++ packages/workflow-core/CHANGELOG.md
@@ -1,9 +1,34 @@
 # @ballerine/workflow-core
 
+## 0.6.44
+
+### Patch Changes
+
+- Bump
+- Updated dependencies
+  - @ballerine/common@0.9.33
+
+## 0.6.43
+
+### Patch Changes
+
+- d
+- Updated dependencies
+  - @ballerine/common@0.9.32
+
+## 0.6.42
+
+### Patch Changes
+
+- version bump
+- Updated dependencies
+  - @ballerine/common@0.9.31
+
 ## 0.6.41
 
 ### Patch Changes
 
+- Bump
 - Updated dependencies
   - @ballerine/common@0.9.30